+++ src/ska_low_mccs_spshw/tile/tile_device.py
@@ -4064,6 +4064,31 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         tango.Except.throw_exception(reason, msg, self.get_name())
         return False
 
+    def _is_synchronised(self: MccsTile, *args: Any) -> bool:
+        """
+        Return a flag representing whether we are in Synchronised state.
+
+        :param args: The tango.AttReqType.
+
+        :return: True if Tile is in Synchronised state.
+        """
+        if self.component_manager._initialise_executing:
+            reason = "CommandNotAllowed"
+            msg = "Cannot execute this command while initialise is executing!"
+            tango.Except.throw_exception(reason, msg, self.get_name())
+            return False
+
+        prog_state = self._attribute_state["tileProgrammingState"].read()[0]
+        if prog_state == "Synchronised":
+            return True
+        reason = "CommandNotAllowed"
+        msg = (
+            "To execute this command we must be in state "
+            f"'Synchronised'! Tile is currently in state {prog_state}"
+        )
+        tango.Except.throw_exception(reason, msg, self.get_name())
+        return False
+
     fpga1Temperature = attribute_from_signal(  # noqa: N815
         fpga1_temperature_signal,
         dtype="DevFloat",
@@ -5862,7 +5887,11 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
             return json.dumps(item_new[0])
         return json.dumps(item_new)
 
-    @command(dtype_in="DevString", dtype_out="DevVarLongStringArray")
+    @command(
+        dtype_in="DevString",
+        dtype_out="DevVarLongStringArray",
+        fisallowed="_is_initialised",
+    )
     @stb.validators.validate_json_args(schema=SetLmcDownload_SCHEMA)
     def SetLmcDownload(
         self: MccsTile,
@@ -5916,7 +5945,11 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
             self.component_manager.refresh_tile_info()
         return (result_codes, messages)
 
-    @command(dtype_in="DevString", dtype_out="DevVarLongStringArray")
+    @command(
+        dtype_in="DevString",
+        dtype_out="DevVarLongStringArray",
+        fisallowed="_is_initialised",
+    )
     @stb.validators.validate_json_args(schema=SetLmcIntegratedDownload_SCHEMA)
     def SetLmcIntegratedDownload(
         self: MccsTile,
@@ -6160,7 +6193,11 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
             cls=NumpyEncoder,
         )
 
-    @command(dtype_in="DevVarLongArray", dtype_out="DevVarLongStringArray")
+    @command(
+        dtype_in="DevVarLongArray",
+        dtype_out="DevVarLongStringArray",
+        fisallowed="_is_initialised",
+    )
     def SetBeamFormerRegions(
         self: MccsTile, argin: list[int]
     ) -> stb.type_hints.DevVarLongStringArrayType:
@@ -6458,7 +6495,11 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         """
         return self.component_manager.apply_calibration(switch_time)
 
-    @command(dtype_in="DevVarDoubleArray", dtype_out="DevVarLongStringArray")
+    @command(
+        dtype_in="DevVarDoubleArray",
+        dtype_out="DevVarLongStringArray",
+        fisallowed="_is_synchronised",
+    )
     def LoadPointingDelays(
         self: MccsTile, argin: list[float]
     ) -> stb.type_hints.DevVarLongStringArrayType:
@@ -6505,7 +6546,11 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
 
         return self.component_manager.load_pointing_delays(delay_array, beam_index)
 
-    @command(dtype_in="DevString", dtype_out="DevVarLongStringArray")
+    @command(
+        dtype_in="DevString",
+        dtype_out="DevVarLongStringArray",
+        fisallowed="_is_synchronised",
+    )
     def ApplyPointingDelays(
         self: MccsTile, start_time: str
     ) -> stb.type_hints.DevVarLongStringArrayType:
@@ -6751,7 +6796,11 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         """
         return self.component_manager.stop_integrated_data()
 
-    @command(dtype_in="DevString", dtype_out="DevVarLongStringArray")
+    @command(
+        dtype_in="DevString",
+        dtype_out="DevVarLongStringArray",
+        fisallowed="_is_synchronised",
+    )
     @stb.validators.validate_json_args(schema=SendDataSamples_SCHEMA)
     def SendDataSamples(
         self: MccsTile,
@@ -7132,7 +7181,7 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         return ([ResultCode.OK], ["StartPatternGenerator command completed OK"])
 
     @engineering_mode_required
-    @command(dtype_out="DevVarLongStringArray")
+    @command(dtype_out="DevVarLongStringArray", fisallowed="_is_synchronised")
     def StartADCs(self: MccsTile) -> stb.type_hints.DevVarLongStringArrayType:
         """
         Start the ADCs.
@@ -7149,7 +7198,7 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         return ([ResultCode.OK], ["StartAdcs command completed OK"])
 
     @engineering_mode_required
-    @command(dtype_out="DevVarLongStringArray")
+    @command(dtype_out="DevVarLongStringArray", fisallowed="_is_synchronised")
     def StopADCs(self: MccsTile) -> stb.type_hints.DevVarLongStringArrayType:
         """
         Stop the ADCs.